5.0.0

+++ webkit/WebCore/page/PerformanceResourceTiming.h
@@ -74,7 +74,7 @@ private:
     PerformanceResourceTiming(const AtomicString& initatorType, const ResourceRequest&, const ResourceResponse&, double initiationTime, double finishTime, Document*);
     ~PerformanceResourceTiming();
 
-    double resourceTimeToDocumentMilliseconds(int deltaMilliseconds) const;
+    double resourceTimeToDocumentMilliseconds(double deltaMilliseconds) const;
 
     AtomicString m_initiatorType;
     ResourceLoadTiming m_timing;
