+++ webkit/WebCore/platform/network/ResourceLoadTiming.h
@@ -87,14 +87,14 @@ public:
     template<class Encoder> void encode(Encoder&) const;
     template<class Decoder> static bool decode(Decoder&, ResourceLoadTiming&);
 
-    // These are millisecond deltas from the navigation start.
-    int domainLookupStart;
-    int domainLookupEnd;
-    int connectStart;
-    int connectEnd;
-    int requestStart;
-    int responseStart;
-    int secureConnectionStart;
+    // These are millisecond deltas from the start time.
+    double domainLookupStart;
+    double domainLookupEnd;
+    double connectStart;
+    double connectEnd;
+    double requestStart;
+    double responseStart;
+    double secureConnectionStart;
 };
 
 template<class Encoder>
