+++ webkit/WebCore/platform/network/curl/ResourceHandleManager.cpp
@@ -152,9 +152,9 @@ static Mutex* sharedResourceMutex(curl_lock_data data) {
 }
 
 #if ENABLE(WEB_TIMING)
-static int milisecondsSinceRequest(double requestTime)
+static double milisecondsSinceRequest(double requestTime)
 {
-    return static_cast<int>((monotonicallyIncreasingTime() - requestTime) * 1000.0);
+    return (monotonicallyIncreasingTime() - requestTime) * 1000.0;
 }
 
 static void calculateWebTimingInformations(ResourceHandleInternal* d)
@@ -172,16 +172,16 @@ static void calculateWebTimingInformations(ResourceHandleInternal* d)
     curl_easy_getinfo(d->m_handle, CURLINFO_PRETRANSFER_TIME, &preTransferTime);
 
     d->m_response.resourceLoadTiming().domainLookupStart = 0;
-    d->m_response.resourceLoadTiming().domainLookupEnd = static_cast<int>(dnslookupTime * 1000);
+    d->m_response.resourceLoadTiming().domainLookupEnd = dnslookupTime * 1000;
 
-    d->m_response.resourceLoadTiming().connectStart = static_cast<int>(dnslookupTime * 1000);
-    d->m_response.resourceLoadTiming().connectEnd = static_cast<int>(connectTime * 1000);
+    d->m_response.resourceLoadTiming().connectStart = dnslookupTime * 1000;
+    d->m_response.resourceLoadTiming().connectEnd = connectTime * 1000;
 
-    d->m_response.resourceLoadTiming().requestStart = static_cast<int>(connectTime *1000);
-    d->m_response.resourceLoadTiming().responseStart =static_cast<int>(preTransferTime * 1000);
+    d->m_response.resourceLoadTiming().requestStart = connectTime *1000;
+    d->m_response.resourceLoadTiming().responseStart = preTransferTime * 1000;
 
     if (appConnectTime)
-        d->m_response.resourceLoadTiming().secureConnectionStart = static_cast<int>(connectTime * 1000);
+        d->m_response.resourceLoadTiming().secureConnectionStart = connectTime * 1000;
 }
 #endif
